(hopefully) made install idempotent

+++ install.sh
@@ -3,20 +3,21 @@
 set -e
 here="$(dirname -- "$(readlink -f -- "$0")")"
 mkdir -p ~/.vim/{autoload,bundle,syntax}
-wget https://github.com/tpope/vim-pathogen/raw/master/autoload/pathogen.vim -O ~/.vim/autoload/pathogen.vim
+[ -f ~/.vim/autoload/pathogen.vim ] || wget https://github.com/tpope/vim-pathogen/raw/master/autoload/pathogen.vim -O ~/.vim/autoload/pathogen.vim
 bash -- "$here/DEPEND"
-ln -s -- "$here/snippets" ~/.vim/UltiSnips
-ln -s -- "$here/snippets" ~/.vim/usnippets
-ln -s -- "$here/genproject" ~/.vim
-ln -s -- "$here/scripts" ~/.vim
+[ -l ~/.vim/UltiSnips ] || ln -s -- "$here/snippets" ~/.vim/UltiSnips
+[ -l ~/.vim/usnippets ] || ln -s -- "$here/snippets" ~/.vim/usnippets
+[ -l ~/.vim/genproject ] || ln -s -- "$here/genproject" ~/.vim/genproject
+[ -l ~/.vim/scripts ] || ln -s -- "$here/scripts" ~/.vim/scripts
 for f in "$here/syntax"/*.vim; do
-	ln -s -- "$f" ~/.vim/syntax
+	[ -l ~/.vim/syntax/"$(basename -- "$f")" ] || ln -s -- "$f" ~/.vim/syntax
 done
 for f in "$here/nerdtree_plugin"/*.vim; do
-	ln -s -- "$f" ~/.vim/bundle/nerdtree/nerdtree_plugin
+	[ -l ~/.vim/bundle/nerdtree/nerdtree_plugin/"$(basename -- "$f")" ] || ln -s -- "$f" ~/.vim/bundle/nerdtree/nerdtree_plugin
 done
-ln -s -- "$here/vimrc" ~/.vim/sniper-vimrc.vim
+[ -l ~/.vim/sniper-vimrc.vim ] || ln -s -- "$here/vimrc" ~/.vim/sniper-vimrc.vim
 touch ~/.vim/site-projects.vim
+if [ "$(test -f ~/.vimrc && head -1 ~/.vimrc)" != "runtime sniper-vimrc.vim" ]; then
 cat >>~/.vimrc <<EOF
 runtime sniper-vimrc.vim
 runtime site-projects.vim
@@ -32,3 +33,4 @@ AddJavaProjectNameStrip j
 "ProvideJavaLibrary jusutils /home/$(whoami)/jutils/dist
 "ProvideJavaLibrary jcmdwin /home/$(whoami)/jcmdwin/dist
 EOF
+fi
